add service file and remove commit
